Using view space for ssao.
Earlier I was using inverse of model view to get to the world space.
Since this ssao expects view space transforming position to view space
worked as expected. Since by definition if a point is occluded it should
have AO == 1.0 and if it is not occluded then it should be 0.0. For lighting
calculations we do

1 - AO